1216 김지민 refactor: 이미지 컴포넌트 위치 이동
